requirements.txt
opencv-python
numpy
pandas
tensorflow
keras
pillow
streamlit
scikit-learn
